requirements.txt
crewai==0.11.2
langchain==0.1.20
langchain_community==0.0.38
langchain_openai==0.0.5
openai==1.3.9
python-dotenv
django
djangorestframework
